Fix Return Type, Add Delete Function, Add Documentation Generation, Fix Module Resolution (#10)
Hi, I realized I made a mistake on the previous PR. TS did not infer the correct type, and I have added it now. This includes a test to show that it is working.

Edit:

I have also added a delete function that maps to the original localStorage delete but triggers a state update as well.

There is now also Github Pages documentation which can be enabled through Travis CI if you set the GITHUB_TOKEN variable in Travis.

Edit Edit:

Fixed a bug with pointing to the wrong "main" file (oops). Now correctly will point to the compiled js location (lib) instead of the index.ts source file.