do not change value if already present in localStorage
Hi, it seems `initialValue` get written to the localStorage on start. This makes using this hook to persist data impossible. Because it overrides existing data at the first `useLocalStorage` call. I've made some changes to prevent existing data to be overridden if key already presents and contains valid json data.